requirements.txt
aniso8601==3.0.0
astroid==1.6.5
autopep8==1.3.5
click==6.7
flasgger==0.9.0
Flask==1.0.2
Flask-API==1.0
Flask-JWT-Extended==3.10.0
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nose==1.3.7
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==1.9.2
pytz==2018.4
PyYAML==3.12
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
